requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=py39h06a4308_0
blas=1.0=mkl
brotli-python=1.0.9=py39h5a03fae_7
bzip2=1.0.8=h7f98852_4
c-ares=1.19.1=h5eee18b_0
ca-certificates=2025.2.25=h06a4308_0
certifi=2025.7.14=py39h06a4308_0
charset-normalizer=3.4.2=pyhd8ed1ab_0
contourpy=1.3.0=pypi_0
cudatoolkit=11.6.0=hecad31d_10
cycler=0.12.1=pypi_0
ffmpeg=4.3=hf484d3e_0
fonttools=4.59.0=pypi_0
freetype=2.10.4=h0708190_1
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
grpcio=1.71.0=py39h6a678d5_0
idna=3.10=pyhd8ed1ab_1
importlib-metadata=8.5.0=py39h06a4308_0
importlib-resources=6.5.2=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
jpeg=9e=h166bdaf_1
kiwisolver=1.4.7=pypi_0
lame=3.100=h7f98852_1001
lcms2=2.11=hcbb858e_1
ld_impl_linux-64=2.40=h12ee557_0
libabseil=20250127.0=cxx17_h6a678d5_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libgrpc=1.71.0=h2d74bed_0
libiconv=1.17=h166bdaf_0
libpng=1.6.37=h21135ba_2
libprotobuf=5.29.3=h3cdef7c_1
libre2-11=2024.07.02=h6a678d5_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.1.0=hc3755c2_3
libxcb=1.17.0=h9b100fa_0
lz4-c=1.9.3=h9c3ff4c_1
markdown=3.8=py39h06a4308_0
markupsafe=3.0.2=py39h5eee18b_0
matplotlib=3.9.4=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7e14d7c_0
mkl_fft=1.3.1=py39h0c7bc48_1
mkl_random=1.2.2=py39hde0f152_0
ncurses=6.5=h7934f7d_0
nettle=3.6=he412f7d_0
numpy=1.24.3=py39h14f4228_0
numpy-base=1.24.3=py39h31eccc5_0
olefile=0.47=pyhd8ed1ab_1
openh264=2.1.1=h780b84a_0
openssl=3.0.17=h5eee18b_0
packaging=25.0=py39h06a4308_0
pandas=2.3.1=pypi_0
pillow=8.1.0=py39h2bb83ca_1
pip=25.1=pyhc872135_2
protobuf=5.29.3=py39he621ea3_0
pthread-stubs=0.3=h0ce48e5_1
pyparsing=3.2.3=pypi_0
pysocks=1.7.1=pyha55dd90_7
python=3.9.20=he870216_1
python-dateutil=2.9.0.post0=pypi_0
python_abi=3.9=2_cp39
pytorch=1.12.1=py3.9_cuda11.6_cudnn8.3.2_0
pytorch-mutex=1.0=cuda
pytorch-wavelets=1.3.0=pypi_0
pytz=2025.2=pypi_0
pywavelets=1.3.0=pypi_0
re2=2024.07.02=hdb19cb5_0
readline=8.2=h5eee18b_0
requests=2.32.4=pyhd8ed1ab_0
scipy=1.13.1=pypi_0
seaborn=0.12.2=pypi_0
setuptools=78.1.1=py39h06a4308_0
six=1.17.0=pyhe01879c_1
sqlite=3.50.2=hb25bd0a_1
tensorboard=2.19.0=py39h06a4308_0
tensorboard-data-server=0.7.0=py39h52d8a92_1
tk=8.6.14=h993c535_1
torchaudio=0.12.1=py39_cu116
torchdiffeq=0.2.2=pypi_0
torchvision=0.13.1=py39_cu116
typing_extensions=4.14.1=pyhe01879c_0
tzdata=2025.2=pypi_0
urllib3=2.2.2=pyhd8ed1ab_0
werkzeug=3.1.3=py39h06a4308_0
wheel=0.45.1=py39h06a4308_0
xorg-libx11=1.8.12=h9b100fa_1
xorg-libxau=1.0.12=h9b100fa_0
xorg-libxdmcp=1.1.5=h9b100fa_0
xorg-xorgproto=2024.1=h5eee18b_1
xz=5.6.4=h5eee18b_1
zipp=3.23.0=pypi_0
zlib=1.2.13=h5eee18b_1
zstd=1.4.9=ha95c52a_0
